refactor: update dir and split dashboard artist
